Move axes_cache to glue.utils.matplotlib
