+++ glue/qt/widgets/scatter_widget.py
@@ -7,13 +7,13 @@ from glue.clients.scatter_client import ScatterClient
 from glue.qt.glue_toolbar import GlueToolbar
 from glue.qt.mouse_mode import (RectangleMode, CircleMode,
                                 PolyMode, HRangeMode, VRangeMode)
-from glue.qt.qtutil import load_ui, cache_axes
+from glue.qt.qtutil import load_ui
 from glue.qt.widget_properties import (ButtonProperty, FloatLineProperty,
                                        CurrentComboProperty,
                                        connect_bool_button, connect_float_edit)
 from glue.qt.widgets.data_viewer import DataViewer
 from glue.qt.widgets.mpl_widget import MplWidget, defer_draw
-from glue.utils import nonpartial
+from glue.utils import nonpartial, cache_axes
 
 __all__ = ['ScatterWidget']
